requirements.txt
streamlit
pandas
pdfminer.six
fpdf
PyPDF2